Changed around the constants for MIN and MAX feedrates. There is a feedrate for XYE and one for Z.
Also, I moved the checking of MIN/MAX to dda.c because getting or not getting a Z move in a G1 code doesn't neccessarily indicate if there is a Z move. I think this change should be ported over the master branch, though not sure who should do it.

All of the changes submitted this weekend have been moved over and tested on my machine :)