Added greycode stepping code
My repstrap electronics have pins that are pulled to ground or +5v to
control the polarity of the coils on each stepper, as opposed to the
mendel/polulu electronics, which have a step/direction pin.  This setup
has to manually keep track of the coil states, and step them in a
greycode sequence to make the steppers move.

I've left the pin definitions as ?_DIR and ?_STEP, but if ?_GREYCODE
is defined it treats them as the two coil polarity pins.

Setting ?_GREYCODE will add 44 bytes per axis, plus 2 bytes of RAM.
Still fits in an ATMega168 though.