Merge woes. For some reason, gcode.c was broken up.
This made my changes wrong, broke all kinds of things. Why was this done again? What benefit is there? Is it really that much harder to end a large file than it is to edit and keep track of two smaller-but-very-closly-related ones?

Whatever. I must admit that I'm mainly bellyaching 'cause of the difficulty it caused me.

I really should at least do a test compile before saying a merge is done.