remove delay between moves.
I can no longer hear a hesitation between moves on the same axis going the same direction.
Fairly minor change, does make the "end interrupt" where it switches take longer, possibly much longer.
This is because it does dda_step at least once, but as many as eight times. Consider the case where
there are a bunch of 0 step "moves" queued. dda_step will run, notice that it didn't move, and mark it as not live.
next_move gets called, and since it's a tail recursive function, it can run through the whole queue in one go.
next_move calls dda_start, which calls dda_step, which may mark it as "not live"
so next_move will notice and call itself again, until queue_empty or dda_start returns with the dda marked as live.
This could possibly take a long time to do, and may skip an interrupt. But it's better than waiting till the next timer tick.
TODO: dda_start shouldn't set the timer if it's unchanged. Right now the steps are in theory slightly farther apart since the timer is reset unconditionally.